+++ test/FASTASpec.hs
@@ -2,10 +2,11 @@
 
 module FASTASpec where
 
-import           Bio.FASTA          (fromFile, toFile)
-import           Bio.FASTA.Type     (FastaItem(..), Fasta)
-import           Bio.Sequence       (bareSequence)
-import           Prelude     hiding (writeFile, readFile)
+import           Bio.FASTA        (fromFile, toFile)
+import           Bio.FASTA.Type   (Fasta, FastaItem (..))
+import           Bio.Sequence     (bareSequence)
+import           Prelude          hiding (readFile, writeFile)
+import           System.Directory (removeFile)
 import           Test.Hspec
 
 correctFasta :: Fasta Char
@@ -27,4 +28,5 @@ writeFile path = describe "writeFile" $ do
     it "correctly write fasta into file" $ do
         toFile correctFasta path
         fasta <- fromFile path
+        removeFile path
         fasta `shouldBe` correctFasta